LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-to-visible-areas-margin-both-expected.txt


FAIL Snap to area such that only the scroll margin from both axes' areas are visible assert_equals: expected 800 but got 500
